# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super TTS daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/tts/super-tts-http.sock` (override with `SUPER_TTS_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/tts/super-tts-http.sock\" \\\n     -H \"Authorization: Bearer $TTS_TOKEN\" \\\n     http://tts.local/v1/ping\n```",
    "contact": {
      "name": "Super TTS",
      "url": "https://github.com/jorge-menjivar/super-tts"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://tts.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/backend/{backend_id}": {
      "delete": {
        "tags": [
          "backends"
        ],
        "summary": "Uninstall a backend",
        "description": "Removes an installed backend and its files from disk. Works for any of them — registry-installed, installed from a custom repo, or imported from a directory.\n\nIf the backend was filling the synthesis stage, that stage is emptied first: the loaded model is unloaded and the selection cleared, so the daemon does not end up pointing at files that no longer exist. `was_active` in the response says whether that happened.\n\nRefused with `409 backend_busy` while an utterance or a streaming session is in flight — removing files out from under one would strand state it still depends on. Errors here use the registry envelope rather than the usual one, because uninstall is the inverse of `POST /registry/backend/install` and a client handles the pair together.",
        "operationId": "uninstall_backend",
        "parameters": [
          {
            "name": "backend_id",
            "in": "path",
            "description": "The backend's id — its `source` as `GET /backend/list` reports it — percent-encoded, e.g. `github.com%2Fsuper-tts%2Fkokoro`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Fsuper-tts%2Fkokoro"
          }
        ],
        "responses": {
          "200": {
            "description": "Removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No installed backend claims that `source` (`not_found`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "409": {
            "description": "An utterance or streaming session is in flight (`backend_busy`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "The files could not be removed (`remove_failed`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/registry/backend/install": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Install a backend",
        "description": "Installs a backend from the registry, from a git-forge repository, or from a directory already staged on this machine — send exactly one of `source`, `repo_url`, or `local_path` (a `repo_url` install also needs `forge`).\n\nThe daemon picks the release asset matching this host's architecture and accelerators, and answers `202` as soon as that choice is made: the download, verification and install run in the background. Follow them on the `registry_install` event topic, keyed by the returned `install_id`. A `warning` of `unverified_source` means the bytes came from somewhere the registry does not vouch for — a custom repo or a local directory — and is worth showing the user.\n\nInstalling neither selects the backend nor loads a model; do that through `POST /pipeline/{stage}` and `POST /pipeline/{stage}/model`. Upgrading one already installed is `POST /registry/backend/update`, which runs this same pipeline.",
        "operationId": "install_registry_backend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted; the install runs in the background.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Not exactly one of `source`, `repo_url`, `local_path` (`bad_request`), a `repo_url` that is not `<host>/<owner>/<repo>` (`bad_repo_url`), or a relative `local_path` (`bad_local_path`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No catalog entry for that `source`, or the repo, release, directory or manifest does not exist (`not_found`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "409": {
            "description": "An install for this backend is already in flight (`install_in_progress`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "422": {
            "description": "No asset matches this host (`incompatible`), or the manifest was rejected (`manifest_invalid`, `manifest_too_large`, `asset_missing`, `source_mismatch`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The forge could not be reached (`forge_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "503": {
            "description": "The catalog could not be fetched and nothing is cached (`registry_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
    "/v1/registry/backend/list": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Browse the published backend catalog",
        "description": "Every backend published to the registry, with the models each serves and whether this machine can run it. `compatibility.compatible` is decided against the host's actual accelerators, so the list reflects what is installable here rather than what exists in general; pass `include_incompatible=true` to see the rest, each with a `compatibility.reason` worth showing the user.\n\n`update_available` is the daemon's answer rather than the client's arithmetic: it compares the installed `backend.toml` on disk against what the index offers, by `source`, and refuses a downgrade. An entry with no `installed_version` is an install, not an update.\n\nThis is what is *available*; `GET /backend/list` is what is installed. The index is cached and re-fetched on a schedule — `POST /registry/backend/refresh` forces that now.",
        "operationId": "list_registry_backends",
        "parameters": [
          {
            "name": "include_incompatible",
            "in": "query",
            "description": "Include entries this machine cannot run. Off by default, so the catalog\nshows what is actually installable here.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by transport — `wasm` or `subprocess`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "online",
            "in": "query",
            "description": "Filter by whether the backend calls out to a network service.",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive substring match over name and description.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The catalog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The catalog could not be fetched and nothing is cached (`registry_unavailable`). Retry, or force a fetch with `POST /registry/backend/refresh`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
    "/v1/registry/backend/refresh": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Re-fetch the backend catalog",
        "description": "Pulls the published index again rather than serving what is cached, and reports how many backends it now holds. Use it after a backend is published, or to clear an `index_stale` marker a listing came back with.\n\n`GET /registry/backend/list` refreshes on its own schedule; this forces it now. Idempotent, and concurrent calls coalesce into a single fetch, so a refresh button cannot start a stampede. The outcome is also published on the `registry_install` event topic, which is how a second client learns the catalog moved.",
        "operationId": "refresh_registry",
        "responses": {
          "200": {
            "description": "Refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The registry could not be reached (`registry_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
    "/v1/registry/backend/update": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Update an installed backend",
        "description": "Upgrades an installed backend to the newest version the catalog offers for this host. Answers with the versions moved between; `noop` is `true` when the installed version was already current, which is a success rather than an error and carries no `install_id`.\n\nOnly a strictly newer semver is an upgrade: a catalog that has gone backwards, or a version neither side can parse, is a no-op rather than a silent downgrade. `update_available` on `GET /registry/backend/list` is the same comparison, so a client can offer the button only where this will do something.\n\nLike install, the work runs in the background — follow `install_id` on the `registry_install` topic when one is returned.",
        "operationId": "update_registry_backend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Already current; nothing was installed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "202": {
            "description": "Accepted; the upgrade runs in the background.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request carried no body (`missing_body`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No catalog entry for that `source` (`not_found`), or nothing is installed under it (`not_installed`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "409": {
            "description": "An install or update for this backend is already in flight (`update_in_progress`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "422": {
            "description": "The newer release has no asset this host can run (`incompatible`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The catalog could not be fetched and nothing is cached (`registry_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "speech_in_progress",
          "download_in_progress",
          "no_switch_in_progress",
          "model_not_loaded",
          "invalid_model",
          "invalid_backend",
          "cuda_unavailable",
          "invalid_device",
          "online_models_disabled",
          "invalid_audio_theme",
          "unsupported_language",
          "invalid_value",
          "not_found",
          "internal",
          "unknown"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "The error envelope, as [`transport.md`] specifies it.\n\n`error_code` is the stable, machine-readable identifier clients switch on\nand the field that determines the HTTP status; `message` is prose for a\nhuman and may be reworded at any time.\n\n[`transport.md`]: https://github.com/jorge-menjivar/super-tts/blob/main/docs/protocol/transport.md",
        "required": [
          "status"
        ],
        "properties": {
          "error_code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "The stable identifier for this failure. Absent only on an unclassified\nserver-side error."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "InstallAccepted": {
        "type": "object",
        "required": [
          "install_id",
          "source",
          "version",
          "selected_asset"
        ],
        "properties": {
          "install_id": {
            "type": "string"
          },
          "selected_asset": {
            "$ref": "#/components/schemas/SelectedAsset"
          },
          "source": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "warning": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "InstallRequest": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "source"
            ],
            "properties": {
              "source": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "repo_url"
            ],
            "properties": {
              "repo_url": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "local_path"
            ],
            "properties": {
              "local_path": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Reason": {
        "type": "object",
        "description": "The `data` object of a [`ReasonEnvelope`].",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Which case of `message` occurred — e.g. `expired`, `exe_changed`,\n`user_denied`."
          }
        }
      },
      "ReasonEnvelope": {
        "type": "object",
        "description": "The auth-failure envelope: `message` names the failure and `data.reason`\nsays which of its cases occurred.\n\nDistinct from [`ErrorEnvelope`] because the auth surface predates\n`error_code` and clients read `data.reason` there. Both are documented\nrather than reconciled, since changing either is a breaking wire change.",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reason"
          },
          "message": {
            "type": "string",
            "description": "The failure identifier, e.g. `invalid_session` or `auth_denied`."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "required": [
          "schema_version",
          "generated_at",
          "backend_count"
        ],
        "properties": {
          "backend_count": {
            "type": "integer",
            "minimum": 0
          },
          "generated_at": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "RegistryError": {
        "type": "object",
        "description": "The registry surface's error envelope, which carries one extra key.\n\nThese endpoints shipped before `error_code` existed, spelling the failure\nidentity as `error`. That key is still sent, because clients read it; the\nstandard `error_code` was added alongside rather than in place of it, so the\nwhole surface honors the \"`error_code` on every error\" rule without breaking\nanyone. Both name the same failure and always agree.\n\nDocumented as its own shape rather than folded into [`ErrorEnvelope`]:\n`error` appears *only* here, and putting it on the shared envelope would\ntell every other endpoint's reader to expect a key they will never receive.",
        "required": [
          "status",
          "error_code",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "The same identifier under the key this surface has always used. Retained\nfor clients written against it; prefer `error_code`.",
            "example": "not_found"
          },
          "error_code": {
            "type": "string",
            "description": "The stable identifier for this failure.",
            "example": "not_found"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail, when there is any to add."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
      "RegistryListResponse": {
        "type": "object",
        "required": [
          "schema_version",
          "generated_at",
          "backends"
        ],
        "properties": {
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistryBackend"
            }
          },
          "generated_at": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "UninstallResponse": {
        "type": "object",
        "description": "What `DELETE /backend/{backend_id}` reports it did.\n\nDeclared beside the handler that answers with it, like every other narrow\nbody on this surface: the shared `super_tts_shared::registry::UninstallResponse`\nis the *client's* deserialization target and says nothing about what the\ndaemon did to get there, while the fields below carry the prose a reader of\nthis endpoint needs — chiefly what `was_active` implies about the stage.\n\nThere is exactly one flag beyond the acknowledgement because super-tts has\nexactly one pipeline stage. A second stage would add a second flag here, not\na second endpoint.",
        "required": [
          "uninstalled",
          "was_active"
        ],
        "properties": {
          "uninstalled": {
            "type": "boolean",
            "description": "Always `true`: the response only exists on the path where the files went\naway. A backend that was never installed is a `404`, not `false`."
          },
          "was_active": {
            "type": "boolean",
            "description": "Whether the removed backend was the one filling the synthesis stage. When\nit was, the stage was emptied before the files went — the loaded model\nunloaded and the selection cleared — so a client showing \"no model\nloaded\" after this is showing the truth rather than a stale read."
          }
        }
      },
...
      "UpdateRequest": {
        "type": "object",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string"
          }
        }
      },
      "UpdateResponse": {
        "type": "object",
        "required": [
          "from_version",
          "to_version",
          "noop"
        ],
        "properties": {
          "from_version": {
            "type": "string"
          },
          "install_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "noop": {
            "type": "boolean"
          },
          "to_version": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}